move constants to core
